pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aes-cbc-decryptor"
version = "1.0.0"
description = "AES-CBC text decryption tool"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Lucas Spawn", email = "lucas@spawn.dev"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Security :: Cryptography",
]
requires-python = ">=3.7"
dependencies = [
    "pycryptodome>=3.15.0",
]

[project.scripts]
aes-decryptor = "aes_decrypt:main"

[project.urls]
Homepage = "https://github.com/SpawNpoa/decifrador"
Repository = "https://github.com/SpawNpoa/decifrador"
Issues = "https://github.com/SpawNpoa/decifrador/issues"

[tool.setuptools]
packages = ["."]

[tool.black]
line-length = 88
target-version = ['py37']

[tool.isort]
profile = "black"
line_length = 88 